* Split seqnn.py into seqnn.py and seqnn_util.py. The latter just contains methods used for various kinds of prediction. These seem to exhibit a large amount of code duplication, and should be condensed in a future CL. In seqnn.py, we now have separate methods for building the predictor, building the loss, etc. All of this supports building the model in terms of data ops that load, for example, from queues, rather than placeholders. The only thing we added to seqnn_util.py is a method for doing prediction when the data processing is in-graph (queues). * Added bin/basenji_train_queues.py This is an end-to-end training/testing example based on bin/basenji_train.py, but that uses data queues rather than placeholders/feed_dicts. The only change in functionality is that we currently don't support any rc/shift data augmentation. This will be added in future work. * Refactored bin/basenji_train.py such that it and bin/basenji_train_queues.py have FLAGs that are defined in bin/shared_flags.py. This facilitates using test files that import both of these files containing main functions.
PiperOrigin-RevId: 169392879
